Updated README with links to deb file Scrubbed all references to Irysview Renamed shell script and systemd files Added checks to ensure script is run as postgres account Added check to ensure BackupDir variable is not empty to enable pg_cleanup function

+++ pgBackup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #########################################################
 #                                                       #
-#       pg_backup wrapper script for BAS hosts          #
+#       pg_backup wrapper script for postgres hosts     #
 #	Please test backups and recoveries!!		#
 #       Prerequisites                                   #
 #          1. Must be run as the postgres user          #
@@ -15,17 +15,18 @@
 #######################pg_back is available for download at #####################################
 #     https://github.com/orgrim/pg_back/releases/download/v2.4.0/pg-back-2.4.0-x86_64.rpm	#
 #################################################################################################
+
 #VariableName
 #_function_name
 
 ############Change to suite your environment#######################
+# Arbitrary name for resulting tar file that contains backup
 DbName=IrysView_Dev
 
 # It is strongly recommend the path below is located on an nfs
 # mount point
 
 BackupPath=/mnt/$(hostname -s)/pgdump 
-
 
 BackupDir=${DbName}-$(date --iso-8601)
 
@@ -34,11 +35,15 @@ _check_prereq () {
 if [[ -f $(which pg_back) && -f $(which pigz) ]]; then
 	return 0
 else
+	echo "Please install pigz and pg-back"
+	echo "yum install pigz -y"
+	echo "yum install https://github.com/orgrim/pg_back/releases/download/v2.4.0/pg-back-2.4.0-x86_64.rpm"
 	return 1
 fi
 }
 
 
+############## Pretty output for journalctl ########################
 _start_msg() {
 cat << EOM
         +-------------------------------------------------------+
@@ -79,9 +84,12 @@ cat << EOM
 EOM
 }
 
+############## End retty output for journalctl #####################
+
 _backup_db () {
-        #pg_back -b "${BackupPath}"/"${BackupDir}" -j 10 -J 2 -Z0 -F directory ${DbName} && \
         pg_back -b "${BackupPath}"/"${BackupDir}" -j 10 -J 2 -Z0 -F directory && \
+	# You can specify a DB Name if you only want to backup 1 DB instead of everything #
+        #pg_back -b "${BackupPath}"/"${BackupDir}" -j 10 -J 2 -Z0 -F directory ${DbName} && \
         _end_msg
         return 0
 }
@@ -106,11 +114,22 @@ _tar_backup () {
 
 _pg_cleanup () {
 # clean up files and directories older than 30 days
-     find ${BackupPath} -maxdepth 1 -name ${DbName}-* -mtime +90 -exec rm -rf "{}" \;
+if [ ! -z ${BackupPath} ]; then
+     	cd ${BackupPath}
+     	find ${BackupPath} -maxdepth 1 -name ${DbName}-* -mtime +90 -exec rm -rf "{}" \;
+else
+	return 1
+fi
+	
 }
 
-#_check_prereq && \
-_start_msg		#start backup message
-_backup_db && \
-_tar_backup
-#_pg_cleanup
+# if [ $(whoami == postgres ];
+# then
+	#_check_prereq && \
+	_start_msg		#start backup message
+	_backup_db && \
+	_tar_backup
+	#_pg_cleanup
+# else
+# echo "This must be run as the postgres user"
+# fi
